src/requirements.txt
python-dateutil
kopf
kubernetes
requests